requirements.txt
flask
flask-cors
flask-jwt-extended
supabase
python-dotenv
bcrypt